% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakery.R
\name{bakery}
\alias{bakery}
\title{Compute robust standard error estimates using multiple sandwich strategies}
\usage{
bakery(
  formula,
  id,
  data,
  family,
  corstr,
  silent = TRUE,
  extraSandwich = "none"
)
}
\arguments{
\item{formula}{A formula expression as for other regression models, of the
form response ~ predictors. See the documentation of `lm` and `formula` for
details.}

\item{id}{A `vector` which identifies the clusters. The length of id should
be the same as the number of observations. Data are assumed to be sorted so
that observations on a cluster are contiguous rows for all entities in the
formula.}

\item{data}{An optional `data frame` in which to interpret the variables
occurring in the formula, along with the id and n variables.}

\item{family}{A `character string` indicating the family for defining link
and variance functions. Currently, only "poisson" is supported.}

\item{corstr}{A `character string` specifying the correlation structure.
Default is "exchangeable". The following are permitted: "independence",
"fixed", "stat_M_dep", "non_stat_M_dep", "exchangeable", "AR-M" and
"unstructured".}

\item{silent}{A `logical` variable controlling whether parameter estimates at
each iteration are printed.}

\item{extraSandwich}{A `character vector` of sandwich estimation procedures
that must be used to compute robust standard errors, additional to the
model standard error and the robust estimator from Liang and Zeger. The
following are permitted: "KC" and "Pan". If not specified or set to "none",
only the model standard error and the robust estimator from Liang and Zeger
will be provided.}
}
\value{
An object of class "gee" including multiple sandwich estimators.
}
\description{
Compute robust standard error estimates using multiple sandwich
  strategies
}
\examples{
## Simulate single gene across 10 patients in 2 groups
n_cells <- 500
gene <- rpois(n_cells, 3)
group_id <- factor(rep(c("control", "treat"), each = n_cells / 2))
patient_id <- factor(rep(paste0("patient", 1:10), each = n_cells / 10))
data <- data.frame(gene, group_id, patient_id)

## Run bakery with KC and Pan extra sandwiches
out <- bakery(
    formula = gene ~ group_id,
    id = "patient_id",
    data = data,
    family = "poisson",
    corstr = "exchangeable",
    silent = TRUE,
    extraSandwich = c("KC", "Pan")
)

## Liang-Zeger variance estimates
diag(out$robust.variance)

## KC variance estimates
diag(out$KC.variance)

## Pan variance estimates
diag(out$Pan.variance)

}
\author{
Jeroen Gilis
}
