% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationsBoxplot.R
\name{correlationsBoxplot}
\alias{correlationsBoxplot}
\title{Visualize inter- and intra-individual correlations by means of boxplots}
\usage{
correlationsBoxplot(object, clustering, assay = NULL)
}
\arguments{
\item{object}{A `SingleCellExperiment` object.}

\item{clustering}{A `character` string indicating from which individual each
cells originates}

\item{assay}{A `character` specifying which `assay` from the
`SingleCellExperiment` to use. If `NULL` will use the first one in
`assayNames(object)`.}
}
\value{
A `ggplot` object.
}
\description{
Visualize inter- and intra-individual correlations by means of
  boxplots
}
\author{
Jeroen Gilis
}
