% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLM.R
\name{fitGLM}
\alias{fitGLM}
\title{Fitting a Generalized Linear Model}
\usage{
fitGLM(
  sce,
  formula,
  family = c("poisson", "quasipoisson"),
  offsets = "TMM",
  ...
)
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} object.}

\item{formula}{An object of class \link[stats:formula]{"formula"} or a character
string that can be coerced to one. Entries in the formula should refer to
columns in \code{colData(sce)}.}

\item{family}{Character string specifying the distribution to be used for
the GLM. Currently supports \code{"poisson"} and \code{"quasipoisson"}.}

\item{offsets}{Either a character string specifying which method to use to
calculate offsets or a numeric vector providing the offsets.}

\item{...}{Further arguments passed on to \code{\link[stats:glm]{stats::glm()}}.}
}
\value{
A list of \code{"glm"} objects for each gene.
}
\description{
Fit a (quasi-)Poisson Generalized Linear Model using \code{\link[stats:glm]{stats::glm()}}, for
further use in DE testing.
}
\details{
The \code{offsets} are calculated using \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}} by default
(\code{method = "TMM"}). Alternatively, a vector of offsets can be provided with
length equal to the number of cells in the data.
}
\examples{
## Mock up data set
sce <- scuttle::mockSCE(ncells = 100, ngenes = 500)
colData(sce)

## Fit model using available colData columns
model_fits <- fitGLM(sce, ~ Mutation_Status + Treatment)
class(model_fits)
length(model_fits)

}
