% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEEutils-package.R
\docType{package}
\name{GEEutils-package}
\alias{GEEutils}
\alias{GEEutils-package}
\title{GEEutils: Aggregation of utility functions for working with Generalized Estimation Equations}
\description{
Aggregation of utility functions for working with Generalized Estimation Equations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jgilis/GEEutils}
}

}
\author{
\strong{Maintainer}: Jeroen Gilis \email{jeroen.gilis@ugent.be}

Authors:
\itemize{
  \item Milan Malfait \email{milan.malfait94@gmail.com} (\href{https://orcid.org/0000-0001-9144-3701}{ORCID})
}

}
\keyword{internal}
