% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeWaldTest.R
\name{geeWaldTest}
\alias{geeWaldTest}
\title{GEE Wald tests}
\usage{
geeWaldTest(object, contrast, SE = "robust.variance")
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object containing fitted
GEE models in the \code{rowData} slot of the object, as obtained with
\code{\link[=bakerySCE]{bakerySCE()}}}

\item{contrast}{A \code{vector} specifying the contrast, i.e. combination of
model parameters of interest.}

\item{SE}{A \verb{character string} that indicates which variance estimator
should be used for inference. Default: \code{"robust.variance"}.}
}
\value{
A \code{data.frame} containing the requested model parameter estimates,
(robust) standard errors, degrees of freedom, Wald test statistics and
p-values.
}
\description{
Perform Wald tests on fitted GEE models from \code{\link[=bakerySCE]{bakerySCE()}}.
}
\examples{
library(scuttle)
set.seed(011235)
sce <- mockSCE(ncells = 400, ngenes = 10)
sce$patient_id <- factor(rep(paste0("patient", 1:8), each = ncol(sce) / 8))
sce$group_id <- factor(rep(paste0("group", 1:2), each = ncol(sce) / 2))

metadata(sce)$formula <- ~ sce$group_id

## Fit models with bakerySCE
sce_fitted <- bakerySCE(
    object = sce,
    id = "patient_id",
    corstr = "independence",
    extraSandwich = "none",
    family = "poisson"
)

## Set up design and contrast
design <- model.matrix(~group_id, data = colData(sce))
L <- matrix(data = 0, ncol = 1, nrow = ncol(design)) # single contrast
rownames(L) <- colnames(design)
L["group_idgroup2", 1] <- 1
colnames(L) <- "group1_vs_group2"

result <- geeWaldTest(object = sce_fitted, contrast = L[, 1], SE = "robust.variance")
## result is a `data.frame` that contains the results of the Wald test for
##  the specified contrasts and working under the robust standard error by
## Liang and Zeger (1986).
}
\seealso{
\code{\link[=bakerySCE]{bakerySCE()}}
}
\author{
Jeroen Gilis, Milan Malfait
}
