% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimulationParameters.R
\name{getSimulationParameters}
\alias{getSimulationParameters}
\title{Obtain parameters for simulating synthetic asRNA-Seq data based on a real
scRNA-Seq dataset.}
\usage{
getSimulationParameters(object, clustering, plot = TRUE, verbose = FALSE)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} instance.}

\item{clustering}{A \verb{character vector} indicating the categorical variable
that is used to color the dimensionality reduction plot. Must be one of
\code{colnames(colData(object))}.}

\item{plot}{\code{Logical} TRUE or FALSE; should intermediate visualizations be
generated?}

\item{verbose}{\code{Logical} TRUE or FALSE; should progress be printed?}
}
\value{
A \code{list} of 10 parameters that can be used to simulate synthetic
scRNA-Seq datasets
}
\description{
Extract parameters from a scRNA-Seq dataset that are relevant
for simulating synthetic asRNA-Seq data.
}
\examples{
sce <- scuttle::mockSCE(ncells = 100, ngenes = 1000)

## Simulate 4 random individuals
sce$patient_id <- factor(rep(paste0("patient", 1:4), each = ncol(sce) / 4))

sim_params <- getSimulationParameters(
    sce, clustering = "patient_id",
    plot = FALSE, verbose = FALSE
)

}
\author{
Jeroen Gilis
}
