% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{glmSandwichTest}
\alias{glmSandwichTest}
\title{GLM Wald tests using sandwich estimators}
\usage{
glmSandwichTest(
  models,
  coef = NULL,
  contrast = NULL,
  subject_id = NULL,
  type = c("HC0", "HC1", "HC2", "HC3", "LiRedden"),
  cadjust = TRUE,
  fix = FALSE
)
}
\arguments{
\item{models}{List of \code{glm} models, resulting from \code{\link[=fitGLM]{fitGLM()}}.}

\item{coef}{Character string or numeric scalar specifying the coefficient to
be tested. Ignored if \code{contrast} is provided.}

\item{contrast}{Integer vector specifying a linear combination of
coefficients to be tested. Should have length equal to the number of
coefficients in the model. Takes precedence over \code{coef}.}

\item{subject_id}{Optional character string specifying a column in the
original data indicating subject IDs. When this is specified, observations
are considered to be correlated within each subject and
\code{\link[sandwich:vcovCL]{sandwich::vcovCL()}} is used to estimate the covariance matrix. When
\code{subject_id = NULL} (the default), a heteroskedastic-consistent sandwich
estimator is used (equivalent to \code{\link[sandwich:vcovHC]{sandwich::vcovHC()}}). Can also provide a
\link[stats:formula]{formula}.}

\item{type}{Character string specifying which type of small-sample bias
adjustment to be applied. Possible values are \code{"HC0"} to \code{"HC3"} or
\code{"LiRedden"}. For details on the \code{"HC*"} adjustments see
\code{\link[sandwich:vcovCL]{sandwich::vcovCL()}}, for \code{"LiRedden"} see "Details" below. The default is
to use \code{"HC0"}, which corresponds to White's estimator.}

\item{cadjust}{Logical, whether to apply cluster bias adjustment. Where
"cluster" refers to the subject-level grouping of the observations. Only
relevant if \code{subject_id} is provided. See \code{\link[sandwich:vcovCL]{sandwich::vcovCL()}} for details.
Default: \code{TRUE}.}

\item{fix}{Logical, whether to fix the covariance matrix to be positive
semi-definite in case it's not. See \code{\link[sandwich:vcovCL]{sandwich::vcovCL()}} for details.
Default: \code{FALSE}.}
}
\value{
A \code{data.frame} with a row for each gene and the following columns:
\itemize{
\item \code{logFC}: log2 Fold-change estimates for the given comparison
\item \code{SE}: standard errors from the requested sandwich estimator type
\item \code{Wald}: Wald statistics
\item \code{PValue}
\item \code{FDR}: FDR-adjusted p-values using Benjamini-Hochberg's method
\item \code{comparison}: the coefficient or contrast tested
}
}
\description{
Perform Wald tests GLM coefficients using sandwich covariance estimators.
When \code{subject_id} is specified, will use the \strong{clustered covariance
estimator}, implemented in \code{\link[sandwich:vcovCL]{sandwich::vcovCL()}}, to account for
within-subject correlations.
}
\details{
The \code{"LiRedden"} adjustment......................
}
\examples{
## Mock up data set
library(scuttle)
sce <- mockSCE(ncells = 100, ngenes = 500)
colData(sce)

## Fit model using available colData columns
model_fits <- fitGLM(sce, ~ Mutation_Status + Treatment)

## Test for Treatment effect using a HC sandwich estimator with HC3 adjustment
res_HC3 <- glmSandwichTest(
    model_fits,
    coef = "Treatmenttreat2",
    type = "HC3"
)
head(res_HC3)

## Test for Treatment effect using a Clustered sandwich estimator with HC3 adjustment
## Assuming cells are correlated within each "Cell_Cycle" level
res_CL <- glmSandwichTest(
    model_fits,
    coef = "Treatmenttreat2",
    subject_id = "Cell_Cycle", type = "HC3"
)
head(res_CL)

}
\references{
Zeileis (2020)
Li & Redden (2015)
}
\seealso{
\code{\link[sandwich:vcovCL]{sandwich::vcovCL()}}
}
