% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationsBoxplot.R
\name{correlationsBoxplot}
\alias{correlationsBoxplot}
\title{Visualize inter- and intra-individual correlations by means of boxplots}
\usage{
correlationsBoxplot(object, clustering, assay = NULL)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{clustering}{A \code{character} string indicating from which individual each
cells originates}

\item{assay}{A \code{character} specifying which \code{assay} from the
\code{SingleCellExperiment} to use. If \code{NULL} will use the first one in
\code{assayNames(object)}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize inter- and intra-individual correlations by means of
boxplots
}
\author{
Jeroen Gilis
}
