% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationsBoxplot.R
\name{correlationsBoxplot}
\alias{correlationsBoxplot}
\title{Visualize inter- and intra-individual correlations by means of boxplots}
\usage{
correlationsBoxplot(object, clustering)
}
\arguments{
\item{object}{A `SingleCellExperiment` object.}

\item{clustering}{A `character` string indicating from which individual each cells originates}
}
\value{
A `ggplot` object.
}
\description{
Visualize inter- and intra-individual correlations by means of boxplots
}
\examples{
TO DO

}
\author{
Jeroen Gilis
}
