% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBySample.R
\name{filterBySample}
\alias{filterBySample}
\title{Function for gene-level filtering.}
\usage{
filterBySample(object, group_id, sample_id, size = 2, assay = NULL)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object}

\item{group_id}{A \code{vector} which identifies the experimental groups of
interest. Must be one of the column names in the \code{colData} of the provided
\code{SingleCellExperiment} object}

\item{sample_id}{A \code{vector} which identifies the samples or experimental
units. Must be one of the column names in the \code{colData} of the provided
\code{SingleCellExperiment} object}

\item{size}{A \code{number} defining the minimum amount of samples that must
express a gene for the gene to pass the filtering. Defaults to 2.}

\item{assay}{A \code{character} specifying which \code{assay} from the
\code{SingleCellExperiment} to use. If \code{NULL} will use the first one in
\code{assayNames(object)}.}
}
\value{
An updated \code{SingleCellExperiment} object, with all genes that did
not pass the filtering removed.
}
\description{
Function for gene-level filtering. Removes genes that are
expressed in less than a certain number of samples (default = 2) in any of
the experimental groups of interest. To be used in combination with
\code{filterByExpr} from the edgeR package.
}
\examples{
library(scuttle)
set.seed(011235)
sce <- mockSCE(ncells = 400, ngenes = 10)
sce$patient_id <- factor(rep(paste0("patient", 1:8), each = ncol(sce) / 8))
sce$group_id <- factor(rep(paste0("group", 1:2), each = ncol(sce) / 2))

metadata(sce)$formula <- ~sce$group_id

sce_filt <- filterBySample(object = sce,
                        group_id = "group_id",
                        sample_id = "patient_id",
                        size = 2,
                        assay = NULL)
sce_filt # filtered SCE object                  

}
\author{
Jeroen Gilis
}
