% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLM.R
\name{fitGLM}
\alias{fitGLM}
\title{Fitting a Generalized Linear Model}
\usage{
fitGLM(
  sce,
  formula,
  family = c("poisson", "quasipoisson"),
  offsets = "TMM",
  ...
)
}
\arguments{
\item{sce}{A \linkS4class{SingleCellExperiment} object.}

\item{formula}{An object of class \link[stats:formula]{"formula"} or a character
string that can be coerced to one. Entries in the formula should refer to
columns in \code{colData(sce)}.}

\item{family}{Character string specifying the distribution to be used for
the GLM. Currently supports \code{"poisson"} and \code{"quasipoisson"}.}

\item{offsets}{Either a character string specifying which method to use to
calculate offsets or a numeric vector providing the offsets.}

\item{...}{Further arguments passed on to \code{\link[stats:glm]{stats::glm()}}.}
}
\value{
A list of \code{"glm"} objects for each gene.
}
\description{
Fitting a Generalized Linear Model
}
\details{
The \code{offsets} ............
}
\examples{

}
