% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeWaldtest.R
\name{geeWaldtest}
\alias{geeWaldtest}
\title{Perform a Wald test on GEE models}
\usage{
geeWaldtest(object, contrast, SE)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object that already contains fitted
GEE models in the \code{rowData} slot of the object, as obtained with the GEEutils
function \code{bakerySCE}.}

\item{contrast}{A \code{matrix} specifying the contrast, i.e. combinations of
model parameters, of interest.}

\item{SE}{A \verb{character string} that indicates}
}
\value{
A \code{Dataframe} containing the requested model parameter estimates,
(robust) standard errors, degrees of freedom, Wald test statistics and
p-values.
}
\description{
Perform a Wald test on GEE models
}
\examples{
library(scuttle)
set.seed(011235)
sce <- mockSCE(ncells = 400, ngenes = 10)
sce$patient_id <- factor(rep(paste0("patient", 1:8), each = ncol(sce) / 8))
sce$group_id <- factor(rep(paste0("group", 1:2), each = ncol(sce) / 2))

sce_fitted <- bakerySCE(object = sce,
                        formula = ~ group_id, 
                        id = "patient_id",
                        corstr = "independence",
                        extraSandwich = "none",
                        family = "poisson")

## Set up design and contrast
design <- model.matrix(~group_id, data = colData(sce))
L <- matrix(data = 0, ncol = 1, nrow = ncol(design)) # single contrast
rownames(L) <- colnames(design)
L["group_idgroup2",1] <- 1
colnames(L) <- "group1_vs_group2"

result <- geeWaldtest(object=sce_fitted, contrast=L[,1], SE="robust.variance")
## result is a `data.frame` that contains the results of the Wald test for
##  the specified contrasts and working under the robust standard error by
## Liang and Zeger (1986).

}
\author{
Jeroen Gilis
}
