% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{glmSandwichTest}
\alias{glmSandwichTest}
\title{GLM Wald tests using sandwich estimators}
\usage{
glmSandwichTest(
  models,
  subject_id,
  coef = NULL,
  contrast = NULL,
  type = c("LiRedden", "HC0", "HC1", "HC2", "HC3"),
  cadjust = FALSE,
  fix = FALSE,
  use_T = FALSE
)
}
\arguments{
\item{models}{List of \code{glm} models, resulting from \code{\link[=fitGLM]{fitGLM()}}.}

\item{subject_id}{Optional character string specifying a column in the
original data indicating subject IDs. When this is specified, observations
are considered to be correlated within each subject and
\code{\link[sandwich:vcovCL]{sandwich::vcovCL()}} is used to estimate the covariance matrix. When
\code{subject_id = NULL} (the default), a heteroskedastic-consistent sandwich
estimator is used (equivalent to \code{\link[sandwich:vcovHC]{sandwich::vcovHC()}}).}

\item{coef}{Character string or numeric scalar specifying the coefficient to
be tested. Ignored if \code{contrast} is provided.}

\item{contrast}{Integer vector specifying a linear combination of
coefficients to be tested. Should have length equal to the number of
coefficients in the model. Takes precedence over \code{coef}.}

\item{type}{Character string specifying which type of small-sample bias
adjustment to be applied. Possible values are \code{"HC0"} to \code{"HC3"} or
\code{"LiRedden"}. For details on the \code{"HC*"} adjustments see
\code{\link[sandwich:vcovCL]{sandwich::vcovCL()}}, for \code{"LiRedden"} see "Details" below. The default is
to use \code{"LiRedden"}.}

\item{cadjust}{Logical, whether to apply cluster bias adjustment. Where
"cluster" refers to the subject-level grouping of the observations. Only
relevant if \code{subject_id} is provided. See \code{\link[sandwich:vcovCL]{sandwich::vcovCL()}} for details.
Default: \code{TRUE}. When using \code{type = "LiRedden"}, this will be set to
\code{FALSE} internally.}

\item{fix}{Logical, whether to fix the covariance matrix to be positive
semi-definite in case it's not. See \code{\link[sandwich:vcovCL]{sandwich::vcovCL()}} for details.
Default: \code{FALSE}. Only relevant for \code{type = "HC2"} or \code{"HC3"}.}

\item{use_T}{Logical, should a \eqn{t}-test be used to calculate the
p-values? If \code{FALSE} (the default), will use a \eqn{z}-test.}
}
\value{
A \code{data.frame} with a row for each gene and the following columns:
\itemize{
\item \code{logFC}: log2 Fold-change estimates for the given comparison
\item \code{SE}: standard errors from the requested sandwich estimator type
\item \code{Wald}: Wald statistics
\item \code{PValue}
\item \code{FDR}: FDR-adjusted p-values using Benjamini-Hochberg's method
\item \code{comparison}: the coefficient or contrast tested
}
}
\description{
Perform Wald tests GLM coefficients using sandwich covariance estimators.
When \code{subject_id} is specified, will use the \strong{clustered covariance
estimator}, implemented in \code{\link[sandwich:vcovCL]{sandwich::vcovCL()}}, to account for
within-subject correlations.
}
\details{
\subsection{Small sample size adjustments}{

Note that 'sample size' here refers to the number of subjects, \emph{not} the
number of cells.

The \code{"LiRedden"} adjustment uses the degrees-of-freedom correction proposed
in Li & Redden (2015). Briefly, the variances of \eqn{\hat{\beta}} are
multiplied with \eqn{K/(K-p)}, where \eqn{K} is the number of subjects and
\eqn{p} is the number of regression parameters.
}
}
\examples{
## Mock up data set
library(scuttle)
sce <- mockSCE(ncells = 100, ngenes = 500)
colData(sce)

## Fit model using available colData columns
model_fits <- fitGLM(sce, ~ Mutation_Status + Treatment)

## Test for Treatment effect using a clustered sandwich estimator with
## Li-Redden adjustment
res <- glmSandwichTest(
    model_fits,
    coef = "Treatmenttreat2",
    subject_id = "Cell_Cycle",
    type = "LiRedden"
)
head(res$table)

}
\references{
Li P. and Redden D. T. 'Small sample performance of bias-corrected sandwich
estimators for cluster-randomized trials with binary outcomes', \emph{Statistics
in Medicine} \strong{34}(2), 281–296 (2015).
doi: \href{https://doi.org/10.1002/sim.6344}{10.1002/sim.6344}.

Zeileis A., Köll S. and Graham N. (2020). 'Various Versatile Variances: An
Object-Oriented Implementation of Clustered Covariances in R', \emph{Journal of
Statistical Software}, \strong{95}(1), 1–36.
doi: \href{https://doi.org/10.18637/jss.v095.i01}{10.18637/jss.v095.i01}.
}
\seealso{
\code{\link[sandwich:vcovCL]{sandwich::vcovCL()}}
}
